feature: Add framework for basic commands
Add a basic structure for (hard-coded) commands.

The basic structure of such a command is defined by the Command
interface, which every individual command needs to implement.

A single command is represented by an object class and gets registered
with the CommandRegistry, that in turn gets queried by the
CommandHandler when necessary.

Currently all commands need to be registered by hard-coding a "put" to
a HashMap in the CommandRegistry, this should get changed in the future,
for example we should just add all object classes implementing the
Command interface.

Issues: HCB-2
Author: Thorben von Rekowski <rekowski.thorben@th0bse.com>