ci: Move pipeline definitions to code
Move pipeline definitions from the web interface to code.
They can now also be tracked in version control, which is convenient
when something goes wrong since we are able to roll back anything that
causes issues with the builds.

Also add a new gradle task to make the version number from the gradle
build script available in TeamCity, so we can use that to tag the
builds.

Issues: #HCB-13 Resolved
Author: Thorben von Rekowski <rekowski.thorben@th0bse.com>