ci: Fix TeamCity cant read build config from VCS
This commit is an attempted fix for the failure of TeamCity to properly
process the build config from settings.kts.

Due to the usage of JDK15 which is more restrictive when it comes to
accessing the internals of Java, the compiler is throwing an error for
the first import statement in the settings.kts.

It should work again if the compiler is told to ignore such warnings,
which is what this commit does.

Issues: #HCB-14 Fixed
Author: Thorben von Rekowski <rekowski.thorben@th0bse.com>