ci: Add build counter back to build identifier
The build counter was not put at the end of the build identifier, which
means that it just consisted of the version number, which doesn't
necessarily change between two builds, which makes it hard to
distinguish between them.

The service message that is sent in the updateTeamCityVersionNumber
gradle task was overwriting the build.number parameter that is
representing the identifier of the build.

The %build.counter% is now added to the end in the service message,
which should fix this issue.

Author: Thorben von Rekowski <rekowski.thorben@th0bse.com>