ci: Fix build id not updated properly
The build identifier didn't get updated properly with the build counter
due to issues with the substitution in the service message that is sent
in the updateTeamCityVersionNumber gradle task.

We now access all configuration parameters that are exposed by TeamCity
and pick the id from there.

Then, the version string from the build.gradle.kts and the build id from
TeamCity get fed back to the build server in the service message as the
identifier for the build.