feat: Add discovery of all command objects
All built-in commands are represented by object classes that implement
the Command interface.

This means that by getting all classes that implement that interface,
the CommandRegistry is able to get all commands that are implemented.

This is working with reflection and then looking for all applicable
classes in the commands sub-package.

Issues: #HCB-10 Fixed
Author: Thorben von Rekowski <rekowski.thorben@th0bse.com>