feat: Add universal data model for storage
To be independent of specific storage solutions (e.g. different
databases), an "intermediate" storage model is needed.

With that, the bot itself doesn't have to deal with any specifics of the
particular storage used and can be sure that every entity (like user or
message) always has the same attributes.

Issues: HCB-8
Author: Thorben von Rekowski <rekowski.thorben@th0bse.com>