feat: Add StorageProvider infrastructure
The StorageProvider Interface serves as a common definition of storage
access, so that the bot doesn't have to care about any implementation
specifics of the particular storage method used.

The specific StorageProvider classes, (currently only the
DatabaseStorageProvider) are then taking care of any specifics and
exposing the storage via the functions defined in the StorageProvider
Interface to the classes that need to interact with the storage.

Issues: HCB-8
Author: Thorben von Rekowski <rekowski.thorben@th0bse.com>